requirements.txt
numpy
pandas
matplotlib
openai
langchain
faiss-cpu
pinecone
streamlit
wandb
python-dotenv
tiktoken
sentence-transformers
tavily-python
ragas
deepeval
datasets
neo4j